backend/requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de datos
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Validación de datos
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Email
sendgrid==6.11.0

# Almacenamiento de archivos
cloudinary==1.36.0

# Utilidades
python-dotenv==1.0.0

# Logging y monitoreo
python-json-logger==2.0.7
